+++ cmf/cmf_core_src/upslope/connections/HBVflow.cpp
@@ -82,13 +82,13 @@ real cmf::upslope::connections::HBVlateral::calc_q(cmf::math::Time t)
 	
 }
 
-void cmf::upslope::connections::HBVlateral::connect_cells( cmf::upslope::Cell & cell1,cmf::upslope::Cell & cell2,int start_at_layer/*=0*/ )
+void cmf::upslope::connections::HBVlateral::connect_cells( cmf::upslope::Cell & cell1,cmf::upslope::Cell & cell2,ptrdiff_t start_at_layer/*=0*/ )
 {
 	real w=cell1.get_topology().flowwidth(cell2);
 	if (w>0)
 	{
-		for (int i = start_at_layer; i < (start_at_layer>=0 ? cell1.layer_count() : 0) ; ++i)	{
-			for (int j = start_at_layer; j < (start_at_layer>=0 ? cell2.layer_count() : 0) ; ++j)	{
+		for (ptrdiff_t i = start_at_layer; i < (start_at_layer>=0 ? cell1.layer_count() : 0) ; ++i)	{
+			for (ptrdiff_t j = start_at_layer; j < (start_at_layer>=0 ? cell2.layer_count() : 0) ; ++j)	{
 				real ca=cell1.get_layer(i)->get_flow_crosssection(*cell2.get_layer(j));
 				if (ca>0)	{
 					real d=cell1.get_layer(i)->position.distanceTo(cell2.get_layer(j)->position);
